Rewrite collectors
They were quite overengineered and also complicated to use. Now, each collector has just one corresponding type, with methods collect_stream() or collect_single(), depending on what you need.

The implementation is much simplified, too. Previously, the collector module was a maze of interconnected macros, generics, GATs. While working on this, I took `current` code as reference because I literally can't find my way around the `next` code.

Previously, each collector type had a method on Context, a method on ShardMessenger, a variant on ShardRunnerMessage, a branch in ShardRunner::handle_filters and a branch in ShardRunner::handle_rx_value. But they all fundamentally do the same thing; so I replaced all of that with a single CollectorCallback system that all collectors use. This reduces the collector-related code in the gateway to like a couple lines of code. And collector.rs itself is delightfully simple too, now.

collector.rs is now a single file, with a collector primitive function called `collect()` which is enough to implement any possible collector. The existing predefined collectors for messages, component interactions, modal interactions, reactions, and generic events still exist though as wrappers for `collect()`. The collector structs are generated using a macro, because their structure is very consistent.

I removed the `removed` filter on ReactionCollector because listening for removed reactions seems very niche, and (like anything else), can be achieved with the `collect()` primitive as well. So ReactionCollector now always collects added reactions. The `collect()` docs show an example of how to collect removed reactions.